+++ src/Config.hs
@@ -33,6 +33,7 @@ import Effect qualified (apply, showDryRun)
 import GHC.Generics (Generic)
 import IO (Time)
 import IO qualified as Time (getCurrent)
+import Opsgenie (Opsgenie, whoIsOnCall)
 import Polysemy (Embed, InterpreterFor, Member, Sem, embed, interpret, makeSem)
 import Polysemy.Error (Error)
 import Polysemy.Log (Log)
@@ -105,17 +106,28 @@ showDryRun resolvedRotationEffectsList =
             & return
         effectLines = padLeft 2 <<$>> Effect.showDryRun members <$> effects
 
-currentResolvedRotationEffects :: (Member Time r) => Conf -> Sem r ResolvedRotationEffects
+currentResolvedRotationEffects ::
+  ( Member Time r,
+    Member Opsgenie r
+  ) =>
+  Conf ->
+  Sem r ResolvedRotationEffects
 currentResolvedRotationEffects conf = do
   userSet <- Set.fromList <$> resolveRotation (rotation conf)
   return (userSet, effects conf)
 
-resolveRotation :: (Member Time r) => Rotation -> Sem r [Text]
+resolveRotation ::
+  ( Member Time r,
+    Member Opsgenie r
+  ) =>
+  Rotation ->
+  Sem r [Text]
 resolveRotation = \case
   Weekly membersList -> do
     time <- Time.getCurrent
     return $ currentCaretaker time <$> membersList
   Const members -> return members
+  OpsgenieScheduleID scheduleID -> whoIsOnCall scheduleID
 
 currentCaretaker :: UTCTime -> [Text] -> Text
 currentCaretaker time candidates = cycle candidates !! utcWeek