+++ ChangeLog.md
@@ -2,6 +2,10 @@
 
 ## Unreleased changes
 
+* Add a OpsgenieScheduleID rotation. The new rotation pulls from Opsgenie the
+  list of emails of whoever is on call currently for the specified schedule.
+  Requires OPSGENIE_API_TOKEN env variable.
+
 ## v0.6.2
 
 * Fix caching in Docker builds.